backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "haha-chatbot-backend"
version = "0.1.0"
description = "Backend service for haha-chatbot"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16,<2",
  "fastapi>=0.115,<1",
  "openai>=1.76,<2",
  "psycopg[binary]>=3.2,<4",
  "pydantic-settings>=2.9,<3",
  "sqlalchemy>=2.0,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
